Rename constants to appConstants
