+++ src/components/SharedComponents/Buttons/Button.js
@@ -1,9 +1,9 @@
 // @flow
 
+import { tailwindFontBold } from "appConstants/fontFamilies.ts";
 import classnames from "classnames";
 import { ActivityIndicator, Heading4, INatIcon } from "components/SharedComponents";
 import { Pressable, View } from "components/styledComponents";
-import { tailwindFontBold } from "constants/fontFamilies.ts";
 import * as React from "react";
 import { useTheme } from "react-native-paper";
